+++ tests/testthat/test-map_snps_to_genes.r
@@ -6,6 +6,7 @@ test_that("map_snps_to_genes works", {
         ctd <- ewceData::ctd()
         #### Run SNP-to-gene mapping ####
         ## Use a very small example GWAS file
+        withr::local_tempdir()
         path_formatted <- MAGMA.Celltyping::get_example_gwas(
             trait = "educational_attainment")
         ## Map
@@ -18,7 +19,7 @@ test_that("map_snps_to_genes works", {
                            "NPARAM","N","ZSTAT","P")
         testthat::expect_true(file.exists(genesOutPath))
         testthat::expect_true(file.exists(gsub(".out",".raw",genesOutPath)))
-        genesOut_dt <- data.table::fread(genesOutPath)
+        genesOut_dt <- data.table::fread(file = genesOutPath)
         testthat::expect_true(all(genesOut_cols %in% colnames(genesOut_dt)))
         testthat::expect_true(nrow(genesOut_dt)>70)