+++ tests/testthat/test-plot_celltype_associations.r
@@ -1,0 +1,12 @@
+test_that("plot_celltype_associations works", {
+  
+    res <- MAGMA.Celltyping::enrichment_results
+    ctAssocs <- res$`finn-a-AD.tsv.gz.35UP.10DOWN`$ctAssocsLinear
+    ctd <- ewceData::ctd()
+    figs <- MAGMA.Celltyping::plot_celltype_associations(
+        ctAssocs = ctAssocs,
+        ctd = ctd)
+    
+    testthat::expect_length(figs,2)
+    testthat::expect_true(methods::is(figs[[1]],"gtable"))
+})